% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{sbeta}
\alias{sbeta}
\title{The Beta Distribution}
\usage{
sbeta(shape1, shape2, statistic = c("all", "mean", "median", "sd", "var",
  "skew", "kurt"))
}
\arguments{
\item{shape1, shape2}{non-negative parameters of the Beta distribution}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the Beta distribution with parameters shape1 and shape2.
}
\examples{
# Jeffreys' Prior
curve(dbeta(x, 0.5, 0.5))
sbeta(0.5, 0.5)

curve(dbeta(x, 5, 1))
sbeta(5, 1)

curve(dbeta(x, 1, 3))
sbeta(1, 3)

curve(dbeta(x, 2, 2))
sbeta(2, 2)

curve(dbeta(x, 2, 5))
sbeta(2, 5)

# Bayesian prior
curve(dbeta(x, 1, 1))
sbeta(1, 1)
}
\seealso{
\code{\link[stats]{Beta}}

    \url{https://en.wikipedia.org/wiki/Beta_distribution}
}
