% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{sbinom}
\alias{sbinom}
\title{The Binomial Distribution}
\usage{
sbinom(size, prob, statistic = c("all", "mean", "median", "sd", "var",
  "skew", "kurt"))
}
\arguments{
\item{size}{number of trials (zero or more).}

\item{prob}{probability of success on each trial.}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the Binomial distribution with parameters size and prob.
}
\examples{
support <- 0:2
plot(support, dbinom(support, 2, 1/2))
sbinom(2, 1/2)

support <- 0:20
plot(support, dbinom(support, 20, 1/2))
sbinom(20, 1/2)

support <- 0:20
plot(support, dbinom(support, 20, 0.7))
sbinom(20, 0.7)

support <- 0:40
plot(support, dbinom(support, 40, 0.5))
sbinom(40, 0.5)
}
\seealso{
\code{\link[stats]{Binomial}}

    \url{https://en.wikipedia.org/wiki/Binomial_distribution}
}
