% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{schisq}
\alias{schisq}
\title{The Chi-Squared Distribution}
\usage{
schisq(df, statistic = c("all", "mean", "median", "sd", "var", "skew",
  "kurt"))
}
\arguments{
\item{df}{degrees of freedom (non-negative, but can be non-integer).}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the chi-squared (chi^2) distribution with df degrees of freedom.
}
\examples{
curve(dchisq(x, 1), xlim = c(0, 8))
schisq(1)

curve(dchisq(x, 2), xlim = c(0, 8))
schisq(2)

curve(dchisq(x, 3), xlim = c(0, 8))
schisq(3)

curve(dchisq(x, 4), xlim = c(0, 8))
schisq(4)

curve(dchisq(x, 6), xlim = c(0, 8))
schisq(6)

curve(dchisq(x, 9), xlim = c(0, 8))
schisq(9)
}
\seealso{
\code{\link[stats]{Chisquare}}

    \url{https://en.wikipedia.org/wiki/Chi-squared_distribution}
}
