% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{sexp}
\alias{sexp}
\title{The Exponential Distribution}
\usage{
sexp(rate = 1, statistic = c("all", "mean", "median", "sd", "var",
  "skew", "kurt"))
}
\arguments{
\item{rate}{rate parameter}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the exponential distribution with rate rate (i.e., mean 1/rate).
}
\examples{
curve(dexp(x, 0.5), xlim = c(0, 5))
sexp(0.5)

curve(dexp(x, 1), xlim = c(0, 5))
sexp()

curve(dexp(x, 1.5), xlim = c(0, 5))
sexp(1.5)
}
\seealso{
\code{\link[stats]{Exponential}}

    \url{https://en.wikipedia.org/wiki/Exponential_distribution}
}
