% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{sgamma}
\alias{sgamma}
\title{The Gamma Distribution}
\usage{
sgamma(shape, rate = 1, statistic = c("all", "mean", "median", "sd",
  "var", "skew", "kurt"))
}
\arguments{
\item{shape}{shape parameter. Must be positive}

\item{rate}{rate parameter. Must be positive}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the Gamma distribution with parameters shape and rate.
}
\examples{
curve(dgamma(x, 1, 1/2), xlim = c(0, 20))
sgamma(1, 1/2)

curve(dgamma(x, 2, 1/2), xlim = c(0, 20))
sgamma(2, 1/2)

curve(dgamma(x, 3, 1/2), xlim = c(0, 20))
sgamma(3, 1/2)

curve(dgamma(x, 5, 1), xlim = c(0, 20))
sgamma(5, 1)

curve(dgamma(x, 9, 2), xlim = c(0, 20))
sgamma(9, 2)

curve(dgamma(x, 7.5, 1), xlim = c(0, 20))
sgamma(7.5, 1)

curve(dgamma(x, 0.5, 1), xlim = c(0, 20))
sgamma(0.5, 1)
}
\seealso{
\code{\link[stats]{GammaDist}}

    \url{https://en.wikipedia.org/wiki/Gamma_distribution}
}
