% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm.R
\name{slnorm}
\alias{slnorm}
\title{The Lognormal Distribution}
\usage{
slnorm(meanlog = 0, sdlog = 1, statistic = c("all", "mean", "median",
  "sd", "var", "skew", "kurt"))
}
\arguments{
\item{meanlog, sdlog}{mean and standard deviation of the distribution on the log scale with default values of 0 and 1 respectively.}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the log normal distribution whose logarithm has mean equal to meanlog and standard deviation equal to sdlog.
}
\examples{
curve(dlnorm(x, 0, 0.25), xlim = c(0, 2.5))
slnorm(0, 0.25)

curve(dlnorm(x, 0, 0.5), xlim = c(0, 2.5))
slnorm(0, 0.5)

curve(dlnorm(x), xlim = c(0, 2.5))
slnorm()
}
\seealso{
\code{\link[stats]{Lognormal}}

    \url{https://en.wikipedia.org/wiki/Log-normal_distribution}
}
