% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{spois}
\alias{spois}
\title{The Poisson Distribution}
\usage{
spois(lambda, statistic = c("all", "mean", "median", "sd", "var", "skew",
  "kurt"))
}
\arguments{
\item{lambda}{mean (non-negative)}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the Poisson distribution with parameter lambda.
}
\examples{
support <- 0:20
plot(support, dpois(support, 1))
spois(1)

support <- 0:20
plot(support, dpois(support, 4))
spois(4)

support <- 0:20
plot(support, dpois(support, 10))
spois(10)
}
\seealso{
\code{\link[stats]{Poisson}}

    \url{https://en.wikipedia.org/wiki/Poisson_distribution}
}
