% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{st}
\alias{st}
\title{The Student t Distribution}
\usage{
st(df, statistic = c("all", "mean", "median", "sd", "var", "skew",
  "kurt"))
}
\arguments{
\item{df}{degrees of freedom (> 0, maybe non-integer). df = Inf is allowed.}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the t distribution with df degrees of freedom.
}
\examples{
curve(dt(x, 1), xlim = c(-5, 5))
st(1)

curve(dt(x, 2), xlim = c(-5, 5))
st(2)

curve(dt(x, 5), xlim = c(-5, 5))
st(5)

curve(dt(x, Inf), xlim = c(-5, 5))
st(Inf)
}
\seealso{
\code{\link[stats]{TDist}}

    \url{https://en.wikipedia.org/wiki/Student's_t-distribution}
}
