% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif.R
\name{sunif}
\alias{sunif}
\title{The Uniform Distribution}
\usage{
sunif(min = 0, max = 1, statistic = c("all", "mean", "median", "sd",
  "var", "skew", "kurt"))
}
\arguments{
\item{min, max}{lower and upper limits of the distribution. Must be finite.}

\item{statistic}{desired descriptive statistic}
}
\value{
named numeric vector
}
\description{
Descriptive statistics for the uniform distribution on the interval from min to max.
}
\examples{
curve(dunif(x))
sunif()
}
\seealso{
\code{\link[stats]{Uniform}}

    \url{https://en.wikipedia.org/wiki/Uniform_distribution_(continuous)}
}
